backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==20.1.0
sqlalchemy==2.1.0
databases==0.7.3
alembic==1.11.1
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib[bcrypt]==1.7.5
aiosmtplib==1.1.6
twilio==8.5.0
joblib==1.3.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
boto3==1.28.57
python-dotenv==1.0.0
slowapi==0.1.4
redis==4.6.0
pytest==7.4.2
pytest-asyncio==0.21.0
httpx==0.24.1